Various Changes.
Implemented support for Area Definitions
    -    Draws colored primitives based on the Area Definition Type
    -    Displays read-only area definition information.
    -    Not editable at the moment. May cause issues due to some data being server side.
    -    Currently cannot see parameters due to limits in Unity's serialization. This will be added in another update.
    -    Can be toggled off using the layer visibility options.
    -    Added area loader window. This can be used to display areas that may not belong to a continent.
    -    Added custom transparent shader.
Updated menu groupings.
Made Zone and Area windows sort by name.
Refactored implementation factories into the ZoneManager.
Added error checking to the asset loading of Forgelight games. This should fix an issue where the progress bar would stop.
Code cleanup.
Implement support for Zone Lights
    -    Area lights functioning as normal, Spotlights require a fix for rotations.
    -    Implemented basic customization, and representation in Unity.
    -    Saved with Zone Objects when exporting zones.
Game Export now only exports LOD0 models.
Major refactor of loading zone data into Unity.
Fixed issue where object ID's were not actually being checked for duplicates on export, and improved performance by using hashsets.
Major refactor of transform conversion between Forgelight and Unity. This is now done mostly by the MathUtils class.

+++ Assets/Scripts/Forgelight/Formats/Zone/Eco.cs
@@ -89,7 +89,7 @@ namespace Forgelight.Formats.Zone
                     byte b = binaryReader.ReadByte();
                     byte a = binaryReader.ReadByte();
 
-                    tint.Color = new Color((float)r/255, (float)g/255, (float)b/255, (float)a/255);
+                    tint.Color = new Color(r/255.0f, g/255.0f, b/255.0f, a/255.0f);
 
                     tint.Percentage = binaryReader.ReadUInt32();
